requirements.txt
faiss-cpu
sentence-transformers
google-generativeai
python-dotenv
pandas
openpyxl
regex
fastapi
uvicorn[standard]
streamlit
requests